Review feedback.
Breaks out the transit IPs/CIDR allowlist functions as requested.
