+++ lib/oxide-vpc/src/engine/gateway/mod.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-// Copyright 2023 Oxide Computer Company
+// Copyright 2024 Oxide Computer Company
 
 //! The Oxide VPC Virtual Gateway.
 //!
@@ -42,7 +42,6 @@
 
 use crate::api::DhcpCfg;
 use crate::api::MacAddr;
-use crate::api::RemoveCidrResp;
 use crate::cfg::Ipv4Cfg;
 use crate::cfg::Ipv6Cfg;
 use crate::cfg::VpcCfg;
@@ -55,8 +54,6 @@ use core::fmt;
 use core::fmt::Display;
 use core::marker::PhantomData;
 use opte::api::Direction;
-use opte::api::IpCidr;
-use opte::api::NoResp;
 use opte::api::OpteError;
 use opte::engine::ether::EtherMod;
 use opte::engine::headers::HeaderAction;
@@ -66,7 +63,6 @@ use opte::engine::layer::LayerActions;
 use opte::engine::packet::InnerFlowId;
 use opte::engine::packet::PacketMeta;
 use opte::engine::port::meta::ActionMeta;
-use opte::engine::port::Port;
 use opte::engine::port::PortBuilder;
 use opte::engine::port::Pos;
 use opte::engine::predicate::DataPredicate;
@@ -76,7 +72,6 @@ use opte::engine::predicate::Ipv6AddrMatch;
 use opte::engine::predicate::Predicate;
 use opte::engine::rule::Action;
 use opte::engine::rule::AllowOrDeny;
-use opte::engine::rule::Finalized;
 use opte::engine::rule::GenHtResult;
 use opte::engine::rule::HdrTransform;
 use opte::engine::rule::MetaAction;
@@ -84,13 +79,13 @@ use opte::engine::rule::ModMetaResult;
 use opte::engine::rule::Rule;
 use opte::engine::rule::StaticAction;
 
-use super::VpcNetwork;
-
 pub mod arp;
 pub mod dhcp;
 pub mod dhcpv6;
 pub mod icmp;
 pub mod icmpv6;
+mod transit;
+pub use transit::*;
 
 pub const NAME: &str = "gateway";
 
@@ -281,88 +276,4 @@ impl Display for VpcMeta {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "vpc-meta")
     }
-}
-
-fn make_holepunch_rules(
-    guest_mac: MacAddr,
-    gateway_mac: MacAddr,
-    dest: IpCidr,
-    vpc_mappings: Arc<VpcMappings>,
-) -> (Rule<Finalized>, Rule<Finalized>) {
-    let vpc_meta = Arc::new(VpcMeta::new(vpc_mappings));
-
-    let (cidr_in_pred, cidr_out_pred) = match dest {
-        IpCidr::Ip4(v4) => (
-            Predicate::InnerDstIp4(vec![Ipv4AddrMatch::Prefix(v4)]),
-            Predicate::InnerSrcIp4(vec![Ipv4AddrMatch::Prefix(v4)]),
-        ),
-        IpCidr::Ip6(v6) => (
-            Predicate::InnerDstIp6(vec![Ipv6AddrMatch::Prefix(v6)]),
-            Predicate::InnerSrcIp6(vec![Ipv6AddrMatch::Prefix(v6)]),
-        ),
-    };
-
-    let mut cidr_out = Rule::new(1000, Action::Meta(vpc_meta));
-    cidr_out.add_predicate(Predicate::InnerEtherSrc(vec![
-        EtherAddrMatch::Exact(guest_mac),
-    ]));
-    cidr_out.add_predicate(cidr_out_pred);
-
-    let mut cidr_in = Rule::new(
-        1000,
-        Action::Static(Arc::new(RewriteSrcMac { gateway_mac })),
-    );
-    cidr_in.add_predicate(cidr_in_pred);
-    cidr_in.add_predicate(Predicate::InnerEtherDst(vec![
-        EtherAddrMatch::Exact(guest_mac),
-    ]));
-
-    (cidr_in.finalize(), cidr_out.finalize())
-}
-
-/// Allows a guest to send and receive traffic on a CIDR block
-/// other than their private IP.
-pub fn allow_cidr(
-    port: &Port<VpcNetwork>,
-    dest: IpCidr,
-    vpc_mappings: Arc<VpcMappings>,
-) -> Result<NoResp, OpteError> {
-    let (in_rule, out_rule) = make_holepunch_rules(
-        port.mac_addr(),
-        port.network().cfg.gateway_mac,
-        dest,
-        vpc_mappings,
-    );
-    port.add_rule(NAME, Direction::In, in_rule)?;
-    port.add_rule(NAME, Direction::Out, out_rule)?;
-    Ok(NoResp::default())
-}
-
-/// Prevents a guest from sending/receiving traffic on a CIDR block
-/// other than their private IP.
-pub fn remove_cidr(
-    port: &Port<VpcNetwork>,
-    dest: IpCidr,
-    vpc_mappings: Arc<VpcMappings>,
-) -> Result<RemoveCidrResp, OpteError> {
-    let (in_rule, out_rule) = make_holepunch_rules(
-        port.mac_addr(),
-        port.network().cfg.gateway_mac,
-        dest,
-        vpc_mappings,
-    );
-    let maybe_in_id = port.find_rule(NAME, Direction::In, &in_rule)?;
-    let maybe_out_id = port.find_rule(NAME, Direction::Out, &out_rule)?;
-    if let Some(id) = maybe_in_id {
-        port.remove_rule(NAME, Direction::In, id)?;
-    }
-    if let Some(id) = maybe_out_id {
-        port.remove_rule(NAME, Direction::Out, id)?;
-    }
-
-    Ok(if maybe_in_id.is_none() || maybe_out_id.is_none() {
-        RemoveCidrResp::NotFound
-    } else {
-        RemoveCidrResp::Ok
-    })
 }
